(PC-4412) call getCurrentPosition without location dialog, without alert
