--- src/libs/geolocation/useGeolocation.d.ts
@@ -1,2 +1,0 @@
-import { GeoCoordinates } from 'react-native-geolocation-service'
-export function useGeolocation(): GeoCoordinates | null
